add JavaDocs and Packages
