+++ src/main/java/Task/Task.java
@@ -1,0 +1,40 @@
+package Task;
+/** 
+ * Represents a <code>task</code> object that contains a string holding the <code>name</code> of the task
+ * and the <code>status</code> of the task.
+ * 
+ * 
+ * @author Brian Quek
+ */
+
+public class Task {
+    private String name;
+    private boolean status;
+
+    /** 
+     * Constructor for the Task object.
+     */
+    public Task(String name) {
+        this.name = name;
+        this.status = false;
+    }
+
+    
+    /** 
+     * Sets the status field based on the input argument.
+     * 
+     * @param status a boolean value that is used to change the status
+     */
+    public void setStatus(boolean status) {
+        this.status = status;
+    }
+    
+    
+    /** 
+     * @return a String containing the status of the task and the task name. 
+     */
+    @Override
+    public String toString() {
+        return String.format("[%s] %s", this.status ? "X" : " ", this.name);
+    }
+}
